requirements.txt
APScheduler==3.6.3
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
decorator==4.4.2
idna==2.10
iniconfig==1.0.1
isort==5.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-telegram-bot==13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
toml==0.10.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.10
wrapt==1.12.1
yapf==0.30.0
